requirements_full.txt
# Twitter/X Automation System Requirements

# Core dependencies
twscrape>=0.12.0
httpx>=0.25.0
asyncio-throttle>=1.0.2

# Telegram bot
python-telegram-bot>=20.0
python-dotenv>=1.0.0

# Data processing
aiofiles>=23.0.0
aiohttp>=3.8.0

# Utilities
fake-useragent>=1.4.0
colorama>=0.4.6

# Optional: for better performance
uvloop>=0.18.0; sys_platform != "win32"

# Development/debugging
loguru>=0.7.0
rich>=13.0.0